Refactor: extract maven processing to its own file
