+++ src/maven/maven-info.js
@@ -1,0 +1,12 @@
+const { default: parse } = require("mvn-artifact-name-parser")
+const { createMavenUrlFromCoordinates } = require("./maven-url")
+
+const generateMavenInfo = async artifact => {
+  const maven = parse(artifact)
+  const mavenUrl = await createMavenUrlFromCoordinates(maven)
+  if (mavenUrl) {
+    maven.url = mavenUrl
+  }
+  return maven
+}
+module.exports = { generateMavenInfo }
